bot/requirements.txt
fastapi
uvicorn
ngrok
python-dotenv
pydantic-settings
httpx
sqlalchemy
psycopg2-binary
pg8000
alembic
redis
celery
google
google-genai
groq